requirements.txt
# python=3.10.19
python-dotenv
modelscope
requests
kaldifeat